Unify utils and move to cogs
