+++ equibot/cogs/__init__.py
@@ -1,4 +1,3 @@
 from .general import General
 from .moderation import Moderation
 from .birthdays import Birthdays
-from .calendarbuider import CalendarBuilder
