+++ equibot/cogs/util/util.py
